Remove duplicate validation in dto and model class
