+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppselectronicmonitoringcreateanorderapi/models/AlcoholMonitoringConditions.kt
@@ -8,9 +8,6 @@ import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.OneToOne
 import jakarta.persistence.Table
-import jakarta.validation.constraints.AssertTrue
-import jakarta.validation.constraints.Future
-import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.data.ValidationErrors
 import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.enums.AlcoholMonitoringType
 import java.time.ZonedDateTime
 import java.util.*
@@ -32,7 +29,6 @@ data class AlcoholMonitoringConditions(
   @Column(name = "START_DATE", nullable = true)
   var startDate: ZonedDateTime? = null,
 
-  @field:Future(message = ValidationErrors.AlcoholMonitoring.END_DATE_MUST_BE_IN_FUTURE)
   @Column(name = "END_DATE", nullable = true)
   var endDate: ZonedDateTime? = null,
 
@@ -40,12 +36,4 @@ data class AlcoholMonitoringConditions(
   @JoinColumn(name = "VERSION_ID", updatable = false, insertable = false)
   private val version: OrderVersion? = null,
 
-) {
-  @AssertTrue(message = ValidationErrors.AlcoholMonitoring.END_DATE_MUST_BE_AFTER_START_DATE)
-  fun isEndDate(): Boolean {
-    if (this.endDate != null && this.startDate != null) {
-      return this.endDate!! > this.startDate
-    }
-    return true
-  }
-}
+)
