+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppselectronicmonitoringcreateanorderapi/models/CurfewConditions.kt
@@ -7,12 +7,7 @@ import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.OneToOne
 import jakarta.persistence.Table
-import jakarta.validation.constraints.AssertTrue
-import jakarta.validation.constraints.Future
-import jakarta.validation.constraints.NotNull
-import jakarta.validation.constraints.Size
 import net.minidev.json.annotate.JsonIgnore
-import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.data.ValidationErrors
 import java.time.ZonedDateTime
 import java.util.*
 
@@ -26,17 +21,13 @@ data class CurfewConditions(
   @Column(name = "VERSION_ID", nullable = false, unique = true)
   val versionId: UUID,
 
-  @field:NotNull(message = ValidationErrors.CurfewConditions.START_DATE_REQUIRED)
   @Column(name = "START_DATE", nullable = true)
   var startDate: ZonedDateTime? = null,
 
-  @field:Future(message = ValidationErrors.CurfewConditions.END_DATE_MUST_BE_IN_FUTURE)
   @Column(name = "END_DATE", nullable = true)
   var endDate: ZonedDateTime? = null,
 
-  @field:NotNull(message = ValidationErrors.CurfewConditions.ADDRESS_REQUIRED)
   @Column(name = "CURFEW_ADDRESS", nullable = true)
-  @field:Size(min = 1, message = ValidationErrors.CurfewConditions.ADDRESS_REQUIRED)
   var curfewAddress: String? = null,
 
   @Column(name = "CURFEW_ADDITIONAL_DETAILS", nullable = true)
@@ -50,12 +41,4 @@ data class CurfewConditions(
   @JoinColumn(name = "VERSION_ID", updatable = false, insertable = false)
   private val version: OrderVersion? = null,
 
-) {
-  @AssertTrue(message = ValidationErrors.CurfewConditions.END_DATE_MUST_BE_AFTER_START_DATE)
-  fun isEndDate(): Boolean {
-    if (this.endDate != null && this.startDate != null) {
-      return this.endDate!! > this.startDate
-    }
-    return true
-  }
-}
+)
