+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppselectronicmonitoringcreateanorderapi/models/CurfewReleaseDateConditions.kt
@@ -9,9 +9,6 @@ import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.OneToOne
 import jakarta.persistence.Table
-import jakarta.validation.constraints.NotNull
-import jakarta.validation.constraints.Size
-import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.data.ValidationErrors
 import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.enums.AddressType
 import java.time.ZonedDateTime
 import java.util.*
@@ -26,22 +23,16 @@ data class CurfewReleaseDateConditions(
   @Column(name = "VERSION_ID", nullable = false, unique = true)
   val versionId: UUID,
 
-  @field:NotNull(message = ValidationErrors.CurfewReleaseDateConditions.START_DATE_REQUIRED)
   @Column(name = "RELEASE_DATE", nullable = true)
   var releaseDate: ZonedDateTime? = null,
 
-  @field:NotNull(message = ValidationErrors.CurfewReleaseDateConditions.START_TIME_REQUIRED)
-  @field:Size(min = 1, message = ValidationErrors.CurfewReleaseDateConditions.START_TIME_REQUIRED)
   @Column(name = "START_TIME", nullable = true)
   var startTime: String? = null,
 
-  @field:NotNull(message = ValidationErrors.CurfewReleaseDateConditions.END_TIME_REQUIRED)
-  @field:Size(min = 1, message = ValidationErrors.CurfewReleaseDateConditions.END_TIME_REQUIRED)
   @Column(name = "END_TIME", nullable = true)
   var endTime: String? = null,
 
   @Enumerated(EnumType.STRING)
-  @field:NotNull(message = ValidationErrors.CurfewReleaseDateConditions.ADDRESS_REQUIRED)
   @Column(name = "CURFEW_ADDRESS", nullable = true)
   var curfewAddress: AddressType? = null,
