+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppselectronicmonitoringcreateanorderapi/models/EnforcementZoneConditions.kt
@@ -9,10 +9,6 @@ import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import jakarta.persistence.Table
-import jakarta.validation.constraints.AssertTrue
-import jakarta.validation.constraints.Future
-import jakarta.validation.constraints.NotNull
-import jakarta.validation.constraints.Size
 import net.minidev.json.annotate.JsonIgnore
 import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.enums.EnforcementZoneType
 import java.time.ZonedDateTime
@@ -30,24 +26,17 @@ data class EnforcementZoneConditions(
 
   @Column(name = "ZONE_TYPE", nullable = false)
   @Enumerated(EnumType.STRING)
-  @field:NotNull(message = "Enforcement zone type is required")
   var zoneType: EnforcementZoneType? = null,
 
-  @field:NotNull(message = "Enforcement zone start date is required")
   @Column(name = "START_DATE", nullable = false)
   var startDate: ZonedDateTime? = null,
 
-  @field:Future(message = "Enforcement zone end date must be in the future")
   @Column(name = "END_DATE", nullable = true)
   var endDate: ZonedDateTime? = null,
 
-  @field:NotNull(message = "Enforcement zone description is required")
-  @field:Size(min = 1, message = "Enforcement zone description is required")
   @Column(name = "DESCRIPTION", nullable = true)
   var description: String? = null,
 
-  @field:NotNull(message = "Enforcement zone duration is required")
-  @field:Size(min = 1, message = "Enforcement zone duration is required")
   @Column(name = "DURATION", nullable = true)
   var duration: String? = null,
 
@@ -67,12 +56,4 @@ data class EnforcementZoneConditions(
   @ManyToOne(optional = true)
   @JoinColumn(name = "VERSION_ID", updatable = false, insertable = false)
   private val version: OrderVersion? = null,
-) {
-  @AssertTrue(message = "End date must be after start date")
-  fun isEndDate(): Boolean {
-    if (this.endDate != null && this.startDate != null) {
-      return this.endDate!! > this.startDate
-    }
-    return true
-  }
-}
+)
