+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppselectronicmonitoringcreateanorderapi/models/dto/UpdateAlcoholMonitoringConditionsDto.kt
@@ -1,5 +1,6 @@
 package uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.dto
 
+import jakarta.validation.constraints.AssertTrue
 import jakarta.validation.constraints.Future
 import jakarta.validation.constraints.NotNull
 import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.data.ValidationErrors
@@ -15,4 +16,12 @@ data class UpdateAlcoholMonitoringConditionsDto(
 
   @field:Future(message = ValidationErrors.AlcoholMonitoring.END_DATE_MUST_BE_IN_FUTURE)
   val endDate: ZonedDateTime? = null,
-)
+) {
+  @AssertTrue(message = ValidationErrors.AlcoholMonitoring.END_DATE_MUST_BE_AFTER_START_DATE)
+  fun isEndDate(): Boolean {
+    if (this.endDate != null && this.startDate != null) {
+      return this.endDate > this.startDate
+    }
+    return true
+  }
+}
