+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppselectronicmonitoringcreateanorderapi/models/dto/UpdateTrailMonitoringConditionsDto.kt
@@ -1,5 +1,6 @@
 package uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.dto
 
+import jakarta.validation.constraints.AssertTrue
 import jakarta.validation.constraints.Future
 import jakarta.validation.constraints.NotNull
 import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.data.ValidationErrors
@@ -11,4 +12,12 @@ data class UpdateTrailMonitoringConditionsDto(
   @field:NotNull(message = ValidationErrors.TrailMonitoringConditions.END_DATE_REQUIRED)
   @field:Future(message = ValidationErrors.TrailMonitoringConditions.END_DATE_MUST_BE_IN_FUTURE)
   val endDate: ZonedDateTime? = null,
-)
+) {
+  @AssertTrue(message = ValidationErrors.TrailMonitoringConditions.END_DATE_MUST_BE_AFTER_START_DATE)
+  fun isEndDate(): Boolean {
+    if (this.endDate != null && this.startDate != null) {
+      return this.endDate!! > this.startDate
+    }
+    return true
+  }
+}
